refactor(auth): remove deprecated authentication classes and consolidate utility functions
