--- src/main/java/com/renewsim/backend/auth/dto/AuthRequestDTO.java
@@ -1,17 +1,0 @@
-package com.renewsim.backend.auth.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-public class AuthRequestDTO {
-    private String username;
-    private String password;
-}
-
-
