--- src/main/java/com/renewsim/backend/security/JwtUtils.java
@@ -1,110 +1,0 @@
-package com.renewsim.backend.security;
-
-import io.jsonwebtoken.*;
-import io.jsonwebtoken.io.Decoders;
-import io.jsonwebtoken.security.Keys;
-import jakarta.annotation.PostConstruct;
-import jakarta.servlet.http.Cookie;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.stereotype.Component;
-
-import java.security.Key;
-import java.util.*;
-
-@Component
-public final class JwtUtils {
-
-    private static final Logger logger = LoggerFactory.getLogger(JwtUtils.class);
-
-    private static final String ROLES_CLAIM = "roles";
-    private static final String SCOPE_CLAIM = "scope";
-    private static final SignatureAlgorithm SIGNATURE_ALGORITHM = SignatureAlgorithm.HS256;
-
-    @Value("${jwt.secret}")
-    private String secret;
-
-    @Value("${jwt.expiration}")
-    private long expiration;
-
-    private Key signingKey;
-
-    @PostConstruct
-    private void init() {
-        this.signingKey = Keys.hmacShaKeyFor(Decoders.BASE64.decode(secret));
-    }
-
-    public String generateToken(String username, Set<String> roles, Set<String> scopes) {
-        return Jwts.builder()
-                .setSubject(username)
-                .claim(ROLES_CLAIM, roles)
-                .claim(SCOPE_CLAIM, String.join(" ", scopes))
-                .setIssuedAt(new Date())
-                .setExpiration(new Date(System.currentTimeMillis() + expiration))
-                .signWith(signingKey, SIGNATURE_ALGORITHM)
-                .compact();
-    }
-
-    public boolean validateToken(String token) {
-        try {
-            parseClaims(token);
-            return true;
-        } catch (JwtException | IllegalArgumentException e) {
-            logger.error("Invalid JWT token: {}", e.getMessage());
-            return false;
-        }
-    }
-
-    public String extractUsername(String token) {
-        return parseClaims(token).getBody().getSubject();
-    }
-
-    @SuppressWarnings("unchecked")
-    public Set<String> extractRoles(String token) {
-        var roles = parseClaims(token).getBody().get(ROLES_CLAIM, List.class);
-        return new HashSet<>(roles);
-    }
-
-    public String extractScopes(String token) {
-        return parseClaims(token).getBody().get(SCOPE_CLAIM, String.class);
-    }
-
-    public String getJwtFromHeader(HttpServletRequest request) {
-        String bearerToken = request.getHeader("Authorization");
-        logger.debug("Authorization Header: {}", bearerToken);
-        if (bearerToken != null && bearerToken.startsWith("Bearer ")) {
-            return bearerToken.substring(7);
-        }
-        return null;
-    }
-
-    private Jws<Claims> parseClaims(String token) {
-        return Jwts.parserBuilder()
-                .setSigningKey(signingKey)
-                .build()
-                .parseClaimsJws(token);
-    }
-
-    public void addJwtCookie(HttpServletResponse response, String token) {
-        Cookie cookie = new Cookie("jwt", token);
-        cookie.setHttpOnly(true);
-        cookie.setSecure(true);
-        cookie.setPath("/");
-        cookie.setMaxAge(7 * 24 * 60 * 60); 
-        response.addCookie(cookie);
-    }
-
-    public void clearJwtCookie(HttpServletResponse response) {
-        Cookie cookie = new Cookie("jwt", null);
-        cookie.setHttpOnly(true);
-        cookie.setSecure(true);
-        cookie.setPath("/");
-        cookie.setMaxAge(0); 
-        response.addCookie(cookie);
-    }
-
-}
