+++ src/test/java/com/renewsim/backend/auth/dto/AuthRequestDTOTest.java
@@ -8,6 +8,8 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
+import com.renewsim.backend.auth.web.dto.AuthRequestDTO;
+
 import java.util.Set;
 
 import static org.assertj.core.api.Assertions.assertThat;
