+++ src/test/java/com/renewsim/backend/auth/dto/AuthResponseDTOTest.java
@@ -3,6 +3,8 @@ package com.renewsim.backend.auth.dto;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
+import com.renewsim.backend.auth.web.dto.AuthResponseDTO;
+
 import java.util.Set;
 
 import static org.assertj.core.api.Assertions.assertThat;
